STRIKE/requirements.txt
altgraph==0.17.4
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
click==8.2.1
colorama==0.4.6
filelock==3.20.0
Flask==3.1.2
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
MouseInfo==0.1.3
mysql-connector==2.2.9
packaging==25.0
pefile==2023.2.7
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
PyMsgBox==2.0.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.2.1
pytweening==1.2.0
pywin32-ctypes==0.2.3
PyYAML==6.0.3
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
Werkzeug==3.1.3